Change the way syntax match is executed.
If the match is executed based on the hard coded color (which should not
be hard coded in the first place, mentioned in issue #1), a color code
can have multiple matches if the hex code is used for multiple colors.

This commit changes the matching as: ' \d+ #[0-9af]+', containing the
xterm256 index of the color, so that the highlight class is independant
from the used RGB code.